Return the appropriate function for reading the specified path/extension

Return the appropriate function for reading the specified path/extension

Usage

file_fn(x, write = FALSE)

Arguments

x

(character) The extension name or the path to the file

write

(logical) Return the writing function? Default FALSE to return the reading function

Value

(function)

See also

Examples

file_fn("csv")
#> function (file, col_names = TRUE, col_types = NULL, col_select = NULL, 
#>     id = NULL, locale = default_locale(), na = c("", "NA"), quoted_na = TRUE, 
#>     quote = "\"", comment = "", trim_ws = TRUE, skip = 0, n_max = Inf, 
#>     guess_max = min(1000, n_max), name_repair = "unique", num_threads = readr_threads(), 
#>     progress = show_progress(), show_col_types = should_show_types(), 
#>     skip_empty_rows = TRUE, lazy = should_read_lazy()) 
#> {
#>     if (edition_first()) {
#>         tokenizer <- tokenizer_csv(na = na, quoted_na = quoted_na, 
#>             quote = quote, comment = comment, trim_ws = trim_ws, 
#>             skip_empty_rows = skip_empty_rows)
#>         return(read_delimited(file, tokenizer, col_names = col_names, 
#>             col_types = col_types, locale = locale, skip = skip, 
#>             skip_empty_rows = skip_empty_rows, comment = comment, 
#>             n_max = n_max, guess_max = guess_max, progress = progress, 
#>             show_col_types = show_col_types))
#>     }
#>     if (!missing(quoted_na)) {
#>         lifecycle::deprecate_soft("2.0.0", "readr::read_csv(quoted_na = )")
#>     }
#>     vroom::vroom(file, delim = ",", col_names = col_names, col_types = col_types, 
#>         col_select = {
#>             {
#>                 col_select
#>             }
#>         }, id = id, .name_repair = name_repair, skip = skip, 
#>         n_max = n_max, na = na, quote = quote, comment = comment, 
#>         skip_empty_rows = skip_empty_rows, trim_ws = trim_ws, 
#>         escape_double = TRUE, escape_backslash = FALSE, locale = locale, 
#>         guess_max = guess_max, show_col_types = show_col_types, 
#>         progress = progress, altrep = lazy, num_threads = num_threads)
#> }
#> <bytecode: 0x161a525a0>
#> <environment: namespace:readr>
file_fn("csv", write = TRUE)
#> function (x, file, na = "NA", append = FALSE, col_names = !append, 
#>     quote = c("needed", "all", "none"), escape = c("double", 
#>         "backslash", "none"), eol = "\n", num_threads = readr_threads(), 
#>     progress = show_progress(), path = deprecated(), quote_escape = deprecated()) 
#> {
#>     if (is_present(path)) {
#>         deprecate_warn("1.4.0", "write_csv(path = )", "write_csv(file = )")
#>         file <- path
#>     }
#>     if (is_present(quote_escape)) {
#>         deprecate_soft("2.0.0", "write_delim(quote_escape = )", 
#>             "write_delim(escape = )")
#>         escape <- quote_escape
#>     }
#>     write_delim(x, file, delim = ",", na = na, append = append, 
#>         col_names = col_names, quote = quote, escape = escape, 
#>         eol = eol, num_threads = num_threads, progress = progress)
#> }
#> <bytecode: 0x136a490e8>
#> <environment: namespace:readr>