Vlookup replace using a lookup column and reference table

Vlookup replace using a lookup column and reference table

Usage

vlookup_from_ref(base, lookup_col, lookup_ref, value_col)

Arguments

base

vector of starting values. Replacements will be made in this vector before it is returned.

lookup_col

vector of values with same length as base that will be matched to lookup_ref to deteremine the replacement indices

lookup_ref

vector of reference values, which lookup_col will be matched to in order to determine replacement values.

value_col

vector of replacement values with same length as lookup_ref

Value

vector

See also

Examples

ref <- tibble::tibble(lookup = letters[1:5], value = 1:5)
original <- tibble::tibble(lookup = letters[1:20], base = runif(20, min = 6, max = 20))
dplyr::mutate(original, base = vlookup_from_ref(base, lookup, ref$lookup, ref$value))
#> # A tibble: 20 × 2
#>    lookup  base
#>    <chr>  <dbl>
#>  1 a       1   
#>  2 b       2   
#>  3 c       3   
#>  4 d       4   
#>  5 e       5   
#>  6 f       6.80
#>  7 g      10.2 
#>  8 h      10.1 
#>  9 i      19.0 
#> 10 j      16.7 
#> 11 k       9.09
#> 12 l       9.60
#> 13 m       9.71
#> 14 n       6.30
#> 15 o       8.48
#> 16 p       9.65
#> 17 q      14.1 
#> 18 r      12.1 
#> 19 s       8.25
#> 20 t       8.87