Day of the week as factor/numeric

Day of the week as factor/numeric

Usage

week_factor(
  x,
  label = FALSE,
  abbr = TRUE,
  week_start = getOption("lubridate.week.start", 7)
)

Arguments

x

chr/Date/POSIXt Days of the week as case insensitive character vector, if using abbreviations, they must be of consistent length. Otherwise, a Date or POSIXt

Value

numeric/factor Depending on the value for label. If x is provided, the day of the week as a factor. If x is not provided, an ordered factor with Monday as 1.

See also

Examples

week_factor()
#>    Sunday    Monday   Tuesday Wednesday  Thursday    Friday  Saturday 
#>       Sun       Mon       Tue       Wed       Thu       Fri       Sat 
#> Levels: Sun < Mon < Tue < Wed < Thu < Fri < Sat
week_factor(c('Tu', 'We'))
#> [1] 3 4
week_factor(c('Tu', 'We'), label = TRUE)
#> [1] Tuesday   Wednesday
#> 7 Levels: Sunday < Monday < Tuesday < Wednesday < Thursday < ... < Saturday
week_factor(seq(lubridate::floor_date(Sys.time(), "year"), Sys.time(), by = "day"), label = TRUE, abbr = FALSE)
#>   [1] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>   [8] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [15] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [22] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [29] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [36] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [43] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [50] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [57] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [64] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [71] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [78] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [85] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [92] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#>  [99] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [106] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [113] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [120] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [127] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [134] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [141] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [148] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [155] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [162] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [169] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [176] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [183] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [190] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [197] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [204] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [211] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [218] Sunday    Monday    Tuesday   Wednesday Thursday  Friday    Saturday 
#> [225] Sunday    Monday    Tuesday   Wednesday Thursday  Friday   
#> 7 Levels: Sunday < Monday < Tuesday < Wednesday < Thursday < ... < Saturday