Get the names of all exported functions in a package

Get the names of all exported functions in a package

Usage

get_package_fns(x, all.names = FALSE, pattern, negate = FALSE)

Arguments

x

(character) Package name

all.names

(logical) Include names that begin with characters . _ etc

Value

(character)

See also

Examples

get_package_fns("dplyr")
#>   [1] "$<-.grouped_df"                                       
#>   [2] "DataMask"                                             
#>   [3] "IS_NUMBER_false"                                      
#>   [4] "IS_NUMBER_oob"                                        
#>   [5] "IS_NUMBER_true"                                       
#>   [6] "Progress"                                             
#>   [7] "[.fun_list"                                           
#>   [8] "[.grouped_df"                                         
#>   [9] "[.rowwise_df"                                         
#>  [10] "[<-.grouped_df"                                       
#>  [11] "[<-.rowwise_df"                                       
#>  [12] "[[<-.grouped_df"                                      
#>  [13] "accumulate"                                           
#>  [14] "accumulate_right"                                     
#>  [15] "across"                                               
#>  [16] "across_glue_mask"                                     
#>  [17] "across_missing_cols_deprecate_warn"                   
#>  [18] "across_setup"                                         
#>  [19] "add_computed_columns"                                 
#>  [20] "add_count"                                            
#>  [21] "add_count.data.frame"                                 
#>  [22] "add_count.default"                                    
#>  [23] "add_count_"                                           
#>  [24] "add_count_impl"                                       
#>  [25] "add_rownames"                                         
#>  [26] "add_suffixes"                                         
#>  [27] "add_tally"                                            
#>  [28] "add_tally_"                                           
#>  [29] "all_equal"                                            
#>  [30] "all_exprs"                                            
#>  [31] "all_vars"                                             
#>  [32] "anti_join"                                            
#>  [33] "anti_join.data.frame"                                 
#>  [34] "any_exprs"                                            
#>  [35] "any_vars"                                             
#>  [36] "append_pos"                                           
#>  [37] "apply_filter_syms"                                    
#>  [38] "apply_unpack_spec"                                    
#>  [39] "arrange"                                              
#>  [40] "arrange.data.frame"                                   
#>  [41] "arrange_"                                             
#>  [42] "arrange_.data.frame"                                  
#>  [43] "arrange_.tbl_df"                                      
#>  [44] "arrange_all"                                          
#>  [45] "arrange_at"                                           
#>  [46] "arrange_if"                                           
#>  [47] "arrange_rows"                                         
#>  [48] "as.data.frame.grouped_df"                             
#>  [49] "as.tbl"                                               
#>  [50] "as.tbl.data.frame"                                    
#>  [51] "as.tbl.tbl"                                           
#>  [52] "as_across_fn_call"                                    
#>  [53] "as_fun"                                               
#>  [54] "as_fun_list"                                          
#>  [55] "as_group_map_function"                                
#>  [56] "as_inlined_function"                                  
#>  [57] "as_join_by"                                           
#>  [58] "as_join_by.character"                                 
#>  [59] "as_join_by.default"                                   
#>  [60] "as_join_by.dplyr_join_by"                             
#>  [61] "as_join_by.list"                                      
#>  [62] "as_pick_expansion"                                    
#>  [63] "as_pick_selection"                                    
#>  [64] "as_tibble.grouped_df"                                 
#>  [65] "as_tibble.rowwise_df"                                 
#>  [66] "auto_by_msg"                                          
#>  [67] "auto_copy"                                            
#>  [68] "auto_copy.data.frame"                                 
#>  [69] "auto_name_formulas"                                   
#>  [70] "bench_tbls"                                           
#>  [71] "between"                                              
#>  [72] "big_mark"                                             
#>  [73] "bind_cols"                                            
#>  [74] "bind_rows"                                            
#>  [75] "binding_join_by_between"                              
#>  [76] "binding_join_by_binary"                               
#>  [77] "binding_join_by_closest"                              
#>  [78] "binding_join_by_dollar"                               
#>  [79] "binding_join_by_equality"                             
#>  [80] "binding_join_by_greater_than"                         
#>  [81] "binding_join_by_greater_than_or_equal"                
#>  [82] "binding_join_by_less_than"                            
#>  [83] "binding_join_by_less_than_or_equal"                   
#>  [84] "binding_join_by_overlaps"                             
#>  [85] "binding_join_by_within"                               
#>  [86] "bullets"                                              
#>  [87] "c_across"                                             
#>  [88] "c_across_missing_cols_deprecate_warn"                 
#>  [89] "c_across_setup"                                       
#>  [90] "can_return_from_exit"                                 
#>  [91] "case_formula_evaluate"                                
#>  [92] "case_match"                                           
#>  [93] "case_when"                                            
#>  [94] "cat_line"                                             
#>  [95] "cbind.grouped_df"                                     
#>  [96] "changes"                                              
#>  [97] "check_arg"                                            
#>  [98] "check_bool"                                           
#>  [99] "check_bounds"                                         
#> [100] "check_by"                                             
#> [101] "check_by_typo"                                        
#> [102] "check_by_typo_impl"                                   
#> [103] "check_call"                                           
#> [104] "check_character"                                      
#> [105] "check_class"                                          
#> [106] "check_closure"                                        
#> [107] "check_compatible"                                     
#> [108] "check_constant"                                       
#> [109] "check_dbplyr"                                         
#> [110] "check_dot_by_typo"                                    
#> [111] "check_dot_cols"                                       
#> [112] "check_duplicate_vars"                                 
#> [113] "check_environment"                                    
#> [114] "check_filter"                                         
#> [115] "check_formula"                                        
#> [116] "check_frac"                                           
#> [117] "check_function"                                       
#> [118] "check_grouped"                                        
#> [119] "check_join_by_dots_empty"                             
#> [120] "check_join_relationship"                              
#> [121] "check_join_vars"                                      
#> [122] "check_keep"                                           
#> [123] "check_length"                                         
#> [124] "check_length_val"                                     
#> [125] "check_logical"                                        
#> [126] "check_missing_arg"                                    
#> [127] "check_muffled_warning"                                
#> [128] "check_n_name"                                         
#> [129] "check_na_matches"                                     
#> [130] "check_name"                                           
#> [131] "check_nth_default"                                    
#> [132] "check_number_decimal"                                 
#> [133] "check_number_whole"                                   
#> [134] "check_size"                                           
#> [135] "check_slice_n_prop"                                   
#> [136] "check_slice_unnamed_n_prop"                           
#> [137] "check_string"                                         
#> [138] "check_symbol"                                         
#> [139] "check_transmute_args"                                 
#> [140] "check_type"                                           
#> [141] "check_unmatched"                                      
#> [142] "clamp"                                                
#> [143] "cli_collapse"                                         
#> [144] "cnd_bullet_combine_details"                           
#> [145] "cnd_bullet_cur_group_label"                           
#> [146] "cnd_bullet_header"                                    
#> [147] "cnd_bullet_rowwise_unlist"                            
#> [148] "cnd_data"                                             
#> [149] "cnd_message_lines"                                    
#> [150] "coalesce"                                             
#> [151] "collapse"                                             
#> [152] "collapse.data.frame"                                  
#> [153] "collect"                                              
#> [154] "collect.data.frame"                                   
#> [155] "combine"                                              
#> [156] "commas"                                               
#> [157] "common_by"                                            
#> [158] "common_by.NULL"                                       
#> [159] "common_by.character"                                  
#> [160] "common_by.default"                                    
#> [161] "common_by.list"                                       
#> [162] "common_by_from_vector"                                
#> [163] "common_handler"                                       
#> [164] "compact"                                              
#> [165] "compact_null"                                         
#> [166] "compare_tbls"                                         
#> [167] "compare_tbls2"                                        
#> [168] "compat_as_lazy"                                       
#> [169] "compat_as_lazy_dots"                                  
#> [170] "compat_lazy"                                          
#> [171] "compat_lazy_dots"                                     
#> [172] "compute"                                              
#> [173] "compute.data.frame"                                   
#> [174] "compute_by"                                           
#> [175] "compute_by_groups"                                    
#> [176] "compute_groups"                                       
#> [177] "compute_join_relationship"                            
#> [178] "consecutive_id"                                       
#> [179] "context_env"                                          
#> [180] "context_local"                                        
#> [181] "context_peek"                                         
#> [182] "context_peek_bare"                                    
#> [183] "context_poke"                                         
#> [184] "copy_to"                                              
#> [185] "copy_to.DBIConnection"                                
#> [186] "copy_to.src_local"                                    
#> [187] "count"                                                
#> [188] "count.data.frame"                                     
#> [189] "count_"                                               
#> [190] "count_regroups"                                       
#> [191] "cross_join"                                           
#> [192] "cross_join.data.frame"                                
#> [193] "ctxt_error_label"                                     
#> [194] "ctxt_error_label_named"                               
#> [195] "cumall"                                               
#> [196] "cumany"                                               
#> [197] "cume_dist"                                            
#> [198] "cummean"                                              
#> [199] "cur_column"                                           
#> [200] "cur_data"                                             
#> [201] "cur_data_all"                                         
#> [202] "cur_group"                                            
#> [203] "cur_group_data"                                       
#> [204] "cur_group_id"                                         
#> [205] "cur_group_label"                                      
#> [206] "cur_group_rows"                                       
#> [207] "current_vars"                                         
#> [208] "data_mask_top"                                        
#> [209] "dataframe_ish"                                        
#> [210] "db_analyze"                                           
#> [211] "db_begin"                                             
#> [212] "db_commit"                                            
#> [213] "db_create_index"                                      
#> [214] "db_create_indexes"                                    
#> [215] "db_create_table"                                      
#> [216] "db_data_type"                                         
#> [217] "db_desc"                                              
#> [218] "db_drop_table"                                        
#> [219] "db_explain"                                           
#> [220] "db_has_table"                                         
#> [221] "db_insert_into"                                       
#> [222] "db_list_tables"                                       
#> [223] "db_query_fields"                                      
#> [224] "db_query_rows"                                        
#> [225] "db_rollback"                                          
#> [226] "db_save_query"                                        
#> [227] "db_write_table"                                       
#> [228] "dense_rank"                                           
#> [229] "deparse_trunc"                                        
#> [230] "desc"                                                 
#> [231] "describe_repair"                                      
#> [232] "detect"                                               
#> [233] "detect_index"                                         
#> [234] "df_unpack"                                            
#> [235] "dim_desc"                                             
#> [236] "discard"                                              
#> [237] "distinct"                                             
#> [238] "distinct.data.frame"                                  
#> [239] "distinct_"                                            
#> [240] "distinct_.data.frame"                                 
#> [241] "distinct_.grouped_df"                                 
#> [242] "distinct_.tbl_df"                                     
#> [243] "distinct_all"                                         
#> [244] "distinct_at"                                          
#> [245] "distinct_if"                                          
#> [246] "distinct_prepare"                                     
#> [247] "do"                                                   
#> [248] "do.NULL"                                              
#> [249] "do.data.frame"                                        
#> [250] "do.grouped_df"                                        
#> [251] "do.rowwise_df"                                        
#> [252] "do_"                                                  
#> [253] "do_.NULL"                                             
#> [254] "do_.data.frame"                                       
#> [255] "do_.grouped_df"                                       
#> [256] "do_.rowwise_df"                                       
#> [257] "dplyr_chop"                                           
#> [258] "dplyr_col_modify"                                     
#> [259] "dplyr_col_modify.data.frame"                          
#> [260] "dplyr_col_modify.grouped_df"                          
#> [261] "dplyr_col_modify.rowwise_df"                          
#> [262] "dplyr_col_select"                                     
#> [263] "dplyr_cumall"                                         
#> [264] "dplyr_cumany"                                         
#> [265] "dplyr_cummean"                                        
#> [266] "dplyr_error_call"                                     
#> [267] "dplyr_error_handler"                                  
#> [268] "dplyr_expand_groups"                                  
#> [269] "dplyr_extract_chunks"                                 
#> [270] "dplyr_group_indices"                                  
#> [271] "dplyr_group_keys"                                     
#> [272] "dplyr_init_library"                                   
#> [273] "dplyr_internal_error"                                 
#> [274] "dplyr_internal_signal"                                
#> [275] "dplyr_lazy_vec_chop_impl"                             
#> [276] "dplyr_legacy_locale"                                  
#> [277] "dplyr_local_error_call"                               
#> [278] "dplyr_local_slice_by_arg"                             
#> [279] "dplyr_locate_matches"                                 
#> [280] "dplyr_locate_sorted_groups"                           
#> [281] "dplyr_make_mask_bindings"                             
#> [282] "dplyr_mask_binding_add"                               
#> [283] "dplyr_mask_binding_remove"                            
#> [284] "dplyr_mask_eval_all"                                  
#> [285] "dplyr_mask_eval_all_filter"                           
#> [286] "dplyr_mask_eval_all_mutate"                           
#> [287] "dplyr_mask_eval_all_summarise"                        
#> [288] "dplyr_new_data_frame"                                 
#> [289] "dplyr_new_list"                                       
#> [290] "dplyr_new_tibble"                                     
#> [291] "dplyr_order_legacy"                                   
#> [292] "dplyr_pick_tibble"                                    
#> [293] "dplyr_proxy_order_legacy"                             
#> [294] "dplyr_quosure_name"                                   
#> [295] "dplyr_quosures"                                       
#> [296] "dplyr_reconstruct"                                    
#> [297] "dplyr_reconstruct.data.frame"                         
#> [298] "dplyr_reconstruct.grouped_df"                         
#> [299] "dplyr_reconstruct.rowwise_df"                         
#> [300] "dplyr_reconstruct_dispatch"                           
#> [301] "dplyr_row_slice"                                      
#> [302] "dplyr_row_slice.data.frame"                           
#> [303] "dplyr_row_slice.grouped_df"                           
#> [304] "dplyr_row_slice.rowwise_df"                           
#> [305] "dplyr_summarise_recycle_chunks_in_place"              
#> [306] "dplyr_validate_grouped_df"                            
#> [307] "dplyr_validate_rowwise_df"                            
#> [308] "dplyr_vec_cast_common"                                
#> [309] "dplyr_vec_ptype_common"                               
#> [310] "dplyr_warning_handler"                                
#> [311] "ensure_group_vars"                                    
#> [312] "env_bind_do_pronouns"                                 
#> [313] "env_inherits_global"                                  
#> [314] "env_resolved"                                         
#> [315] "equal_data_frame"                                     
#> [316] "err_expr"                                             
#> [317] "err_locs"                                             
#> [318] "err_vars"                                             
#> [319] "error_call"                                           
#> [320] "error_label"                                          
#> [321] "error_label_named"                                    
#> [322] "error_name_length_required"                           
#> [323] "eval_join_by_between"                                 
#> [324] "eval_join_by_binary"                                  
#> [325] "eval_join_by_closest"                                 
#> [326] "eval_join_by_dollar"                                  
#> [327] "eval_join_by_overlaps"                                
#> [328] "eval_join_by_within"                                  
#> [329] "eval_pick"                                            
#> [330] "eval_relocate"                                        
#> [331] "eval_select_by"                                       
#> [332] "eval_tbls"                                            
#> [333] "eval_tbls2"                                           
#> [334] "every"                                                
#> [335] "expand_across"                                        
#> [336] "expand_groups"                                        
#> [337] "expand_if_across"                                     
#> [338] "expand_pick"                                          
#> [339] "expand_pick_call"                                     
#> [340] "expand_pick_quo"                                      
#> [341] "explain"                                              
#> [342] "expr_substitute"                                      
#> [343] "failwith"                                             
#> [344] "ffi_dplyr_reconstruct"                                
#> [345] "ffi_test_dplyr_attributes"                            
#> [346] "ffi_test_dplyr_set_attributes"                        
#> [347] "filter"                                               
#> [348] "filter.data.frame"                                    
#> [349] "filter.ts"                                            
#> [350] "filter_"                                              
#> [351] "filter_.data.frame"                                   
#> [352] "filter_.tbl_df"                                       
#> [353] "filter_all"                                           
#> [354] "filter_at"                                            
#> [355] "filter_bullets"                                       
#> [356] "filter_bullets.dplyr:::filter_incompatible_size"      
#> [357] "filter_bullets.dplyr:::filter_incompatible_type"      
#> [358] "filter_eval"                                          
#> [359] "filter_expand"                                        
#> [360] "filter_if"                                            
#> [361] "filter_rows"                                          
#> [362] "finalise_equi_join_by"                                
#> [363] "find_template"                                        
#> [364] "find_var"                                             
#> [365] "first"                                                
#> [366] "flat_map_chr"                                         
#> [367] "fmt_check_length_val"                                 
#> [368] "fmt_classes"                                          
#> [369] "fmt_cols"                                             
#> [370] "fmt_comma"                                            
#> [371] "fmt_obj"                                              
#> [372] "fmt_obj1"                                             
#> [373] "fmt_pos_args"                                         
#> [374] "format.src_local"                                     
#> [375] "from_testthat"                                        
#> [376] "full_join"                                            
#> [377] "full_join.data.frame"                                 
#> [378] "fun_env"                                              
#> [379] "funs"                                                 
#> [380] "funs_"                                                
#> [381] "get_slice_size"                                       
#> [382] "group_by"                                             
#> [383] "group_by.data.frame"                                  
#> [384] "group_by_"                                            
#> [385] "group_by_.data.frame"                                 
#> [386] "group_by_.rowwise_df"                                 
#> [387] "group_by_all"                                         
#> [388] "group_by_at"                                          
#> [389] "group_by_drop_default"                                
#> [390] "group_by_drop_default.default"                        
#> [391] "group_by_drop_default.grouped_df"                     
#> [392] "group_by_if"                                          
#> [393] "group_by_prepare"                                     
#> [394] "group_cols"                                           
#> [395] "group_cols_legacy"                                    
#> [396] "group_data"                                           
#> [397] "group_data.data.frame"                                
#> [398] "group_data.grouped_df"                                
#> [399] "group_data.rowwise_df"                                
#> [400] "group_data.tbl_df"                                    
#> [401] "group_data_trim"                                      
#> [402] "group_indices"                                        
#> [403] "group_indices.data.frame"                             
#> [404] "group_indices_"                                       
#> [405] "group_indices_.data.frame"                            
#> [406] "group_indices_.grouped_df"                            
#> [407] "group_indices_.rowwise_df"                            
#> [408] "group_intersect"                                      
#> [409] "group_keys"                                           
#> [410] "group_keys.data.frame"                                
#> [411] "group_keys0"                                          
#> [412] "group_labels_details"                                 
#> [413] "group_map"                                            
#> [414] "group_map.data.frame"                                 
#> [415] "group_modify"                                         
#> [416] "group_modify.data.frame"                              
#> [417] "group_modify.grouped_df"                              
#> [418] "group_nest"                                           
#> [419] "group_nest.data.frame"                                
#> [420] "group_nest.grouped_df"                                
#> [421] "group_nest_impl"                                      
#> [422] "group_rows"                                           
#> [423] "group_size"                                           
#> [424] "group_size.data.frame"                                
#> [425] "group_split"                                          
#> [426] "group_split.data.frame"                               
#> [427] "group_split.grouped_df"                               
#> [428] "group_split.rowwise_df"                               
#> [429] "group_split_impl"                                     
#> [430] "group_sum"                                            
#> [431] "group_trim"                                           
#> [432] "group_trim.data.frame"                                
#> [433] "group_trim.grouped_df"                                
#> [434] "group_vars"                                           
#> [435] "group_vars.data.frame"                                
#> [436] "group_walk"                                           
#> [437] "grouped_df"                                           
#> [438] "groups"                                               
#> [439] "groups.data.frame"                                    
#> [440] "has_active_group_context"                             
#> [441] "has_minimum_stringi"                                  
#> [442] "help_topic"                                           
#> [443] "id"                                                   
#> [444] "ident"                                                
#> [445] "if_across"                                            
#> [446] "if_all"                                               
#> [447] "if_any"                                               
#> [448] "if_else"                                              
#> [449] "imap"                                                 
#> [450] "import_from"                                          
#> [451] "import_vctrs"                                         
#> [452] "inner_join"                                           
#> [453] "inner_join.data.frame"                                
#> [454] "intersect.data.frame"                                 
#> [455] "is.grouped_df"                                        
#> [456] "is.src"                                               
#> [457] "is.tbl"                                               
#> [458] "is_compatible"                                        
#> [459] "is_cross_by"                                          
#> [460] "is_data_pronoun"                                      
#> [461] "is_direct"                                            
#> [462] "is_fun_list"                                          
#> [463] "is_grouped_df"                                        
#> [464] "is_inlinable_formula"                                 
#> [465] "is_inlinable_function"                                
#> [466] "is_inlinable_lambda"                                  
#> [467] "is_rowwise_df"                                        
#> [468] "is_sel_vars"                                          
#> [469] "is_slice_call"                                        
#> [470] "is_symbol_or_string"                                  
#> [471] "is_syntactic"                                         
#> [472] "join_by"                                              
#> [473] "join_by_common"                                       
#> [474] "join_cast_common"                                     
#> [475] "join_cols"                                            
#> [476] "join_filter"                                          
#> [477] "join_mutate"                                          
#> [478] "join_rows"                                            
#> [479] "keep"                                                 
#> [480] "label_output_dataframe"                               
#> [481] "label_output_list"                                    
#> [482] "lag"                                                  
#> [483] "last"                                                 
#> [484] "last_dplyr_warnings"                                  
#> [485] "lazy_deprec"                                          
#> [486] "lead"                                                 
#> [487] "left_join"                                            
#> [488] "left_join.data.frame"                                 
#> [489] "list_flatten"                                         
#> [490] "local_column"                                         
#> [491] "local_error_context"                                  
#> [492] "local_mask"                                           
#> [493] "locale_to_chr_proxy_collate"                          
#> [494] "location"                                             
#> [495] "make_syntactic"                                       
#> [496] "make_tbl"                                             
#> [497] "manip_all"                                            
#> [498] "manip_apply_syms"                                     
#> [499] "manip_at"                                             
#> [500] "manip_if"                                             
#> [501] "map"                                                  
#> [502] "map2"                                                 
#> [503] "map2_chr"                                             
#> [504] "map2_dbl"                                             
#> [505] "map2_int"                                             
#> [506] "map2_lgl"                                             
#> [507] "map_chr"                                              
#> [508] "map_dbl"                                              
#> [509] "map_if"                                               
#> [510] "map_int"                                              
#> [511] "map_lgl"                                              
#> [512] "mask_type"                                            
#> [513] "maybe_restart"                                        
#> [514] "methods_generic"                                      
#> [515] "methods_rd"                                           
#> [516] "min_rank"                                             
#> [517] "minimal_names"                                        
#> [518] "mutate"                                               
#> [519] "mutate.data.frame"                                    
#> [520] "mutate_"                                              
#> [521] "mutate_.data.frame"                                   
#> [522] "mutate_.tbl_df"                                       
#> [523] "mutate_all"                                           
#> [524] "mutate_at"                                            
#> [525] "mutate_bullets"                                       
#> [526] "mutate_bullets.dplyr:::error_incompatible_combine"    
#> [527] "mutate_bullets.dplyr:::mutate_constant_recycle_error" 
#> [528] "mutate_bullets.dplyr:::mutate_incompatible_size"      
#> [529] "mutate_bullets.dplyr:::mutate_mixed_null"             
#> [530] "mutate_bullets.dplyr:::mutate_not_vector"             
#> [531] "mutate_col"                                           
#> [532] "mutate_cols"                                          
#> [533] "mutate_each"                                          
#> [534] "mutate_each_"                                         
#> [535] "mutate_each_impl"                                     
#> [536] "mutate_if"                                            
#> [537] "mutate_keep"                                          
#> [538] "mutate_relocate"                                      
#> [539] "n"                                                    
#> [540] "n_distinct"                                           
#> [541] "n_groups"                                             
#> [542] "n_groups.data.frame"                                  
#> [543] "n_name"                                               
#> [544] "na_if"                                                
#> [545] "named_args"                                           
#> [546] "names<-.grouped_df"                                   
#> [547] "names<-.rowwise_df"                                   
#> [548] "names_as_error_names"                                 
#> [549] "near"                                                 
#> [550] "negate"                                               
#> [551] "nest_by"                                              
#> [552] "nest_by.data.frame"                                   
#> [553] "nest_by.grouped_df"                                   
#> [554] "nest_join"                                            
#> [555] "nest_join.data.frame"                                 
#> [556] "new_by"                                               
#> [557] "new_dplyr_quosure"                                    
#> [558] "new_dplyr_warning"                                    
#> [559] "new_error_context"                                    
#> [560] "new_expanded_quosures"                                
#> [561] "new_funs"                                             
#> [562] "new_grouped_df"                                       
#> [563] "new_join_by"                                          
#> [564] "new_rowwise_df"                                       
#> [565] "new_sel_vars"                                         
#> [566] "node_walk_replace"                                    
#> [567] "now"                                                  
#> [568] "ntext"                                                
#> [569] "nth"                                                  
#> [570] "ntile"                                                
#> [571] "obj_type_friendly"                                    
#> [572] "obj_type_oo"                                          
#> [573] "or_1"                                                 
#> [574] "order_by"                                             
#> [575] "oxford_comma"                                         
#> [576] "parse_join_by_between"                                
#> [577] "parse_join_by_binary"                                 
#> [578] "parse_join_by_closest"                                
#> [579] "parse_join_by_dollar"                                 
#> [580] "parse_join_by_expr"                                   
#> [581] "parse_join_by_name"                                   
#> [582] "parse_join_by_overlaps"                               
#> [583] "parse_join_by_within"                                 
#> [584] "paste_line"                                           
#> [585] "peek_column"                                          
#> [586] "peek_error_context"                                   
#> [587] "peek_mask"                                            
#> [588] "percent_rank"                                         
#> [589] "pick"                                                 
#> [590] "pmap"                                                 
#> [591] "poke_error_context"                                   
#> [592] "print.all_vars"                                       
#> [593] "print.any_vars"                                       
#> [594] "print.dplyr_join_by"                                  
#> [595] "print.dplyr_sel_vars"                                 
#> [596] "print.fun_list"                                       
#> [597] "print.last_dplyr_warnings"                            
#> [598] "print.src"                                            
#> [599] "progress_estimated"                                   
#> [600] "pull"                                                 
#> [601] "pull.data.frame"                                      
#> [602] "push_dplyr_warnings"                                  
#> [603] "quo_as_function"                                      
#> [604] "quo_as_label"                                         
#> [605] "quo_eval_fns"                                         
#> [606] "quo_is_variable_reference"                            
#> [607] "quo_reduce"                                           
#> [608] "quo_set_env_to_data_mask_top"                         
#> [609] "random_table_name"                                    
#> [610] "rbind.grouped_df"                                     
#> [611] "re_match"                                             
#> [612] "recode"                                               
#> [613] "recode.character"                                     
#> [614] "recode.factor"                                        
#> [615] "recode.numeric"                                       
#> [616] "recode_default"                                       
#> [617] "recode_default.default"                               
#> [618] "recode_default.factor"                                
#> [619] "recode_factor"                                        
#> [620] "reduce"                                               
#> [621] "reduce_right"                                         
#> [622] "reframe"                                              
#> [623] "reframe.data.frame"                                   
#> [624] "relocate"                                             
#> [625] "relocate.data.frame"                                  
#> [626] "rename"                                               
#> [627] "rename.data.frame"                                    
#> [628] "rename_"                                              
#> [629] "rename_.data.frame"                                   
#> [630] "rename_.grouped_df"                                   
#> [631] "rename_all"                                           
#> [632] "rename_at"                                            
#> [633] "rename_if"                                            
#> [634] "rename_vars"                                          
#> [635] "rename_vars_"                                         
#> [636] "rename_with"                                          
#> [637] "rename_with.data.frame"                               
#> [638] "replace_with"                                         
#> [639] "reset_dplyr_warnings"                                 
#> [640] "rethrow_error_join_incompatible_type"                 
#> [641] "rethrow_error_join_matches_incomplete"                
#> [642] "rethrow_error_join_matches_multiple"                  
#> [643] "rethrow_error_join_matches_nothing"                   
#> [644] "rethrow_error_join_matches_overflow"                  
#> [645] "rethrow_error_join_matches_remaining"                 
#> [646] "rethrow_error_join_relationship_many_to_one"          
#> [647] "rethrow_error_join_relationship_one_to_many"          
#> [648] "rethrow_error_join_relationship_one_to_one"           
#> [649] "rethrow_warning_join_matches_multiple"                
#> [650] "rethrow_warning_join_relationship_many_to_many"       
#> [651] "right_join"                                           
#> [652] "right_join.data.frame"                                
#> [653] "row_number"                                           
#> [654] "rows_append"                                          
#> [655] "rows_append.data.frame"                               
#> [656] "rows_bind"                                            
#> [657] "rows_cast_y"                                          
#> [658] "rows_check_by"                                        
#> [659] "rows_check_conflict"                                  
#> [660] "rows_check_contains_by"                               
#> [661] "rows_check_unique"                                    
#> [662] "rows_check_unmatched"                                 
#> [663] "rows_check_x_contains_y"                              
#> [664] "rows_check_y_conflict"                                
#> [665] "rows_check_y_unmatched"                               
#> [666] "rows_delete"                                          
#> [667] "rows_delete.data.frame"                               
#> [668] "rows_df_in_place"                                     
#> [669] "rows_insert"                                          
#> [670] "rows_insert.data.frame"                               
#> [671] "rows_patch"                                           
#> [672] "rows_patch.data.frame"                                
#> [673] "rows_update"                                          
#> [674] "rows_update.data.frame"                               
#> [675] "rows_upsert"                                          
#> [676] "rows_upsert.data.frame"                               
#> [677] "rowwise"                                              
#> [678] "rowwise.data.frame"                                   
#> [679] "rowwise.grouped_df"                                   
#> [680] "rowwise_df"                                           
#> [681] "rule"                                                 
#> [682] "same_src"                                             
#> [683] "same_src.data.frame"                                  
#> [684] "sample_frac"                                          
#> [685] "sample_frac.data.frame"                               
#> [686] "sample_frac.default"                                  
#> [687] "sample_int"                                           
#> [688] "sample_n"                                             
#> [689] "sample_n.data.frame"                                  
#> [690] "sample_n.default"                                     
#> [691] "select"                                               
#> [692] "select.data.frame"                                    
#> [693] "select.list"                                          
#> [694] "select_"                                              
#> [695] "select_.data.frame"                                   
#> [696] "select_.grouped_df"                                   
#> [697] "select_all"                                           
#> [698] "select_at"                                            
#> [699] "select_if"                                            
#> [700] "select_var"                                           
#> [701] "select_vars"                                          
#> [702] "select_vars_"                                         
#> [703] "semi_join"                                            
#> [704] "semi_join.data.frame"                                 
#> [705] "set_minimal_names"                                    
#> [706] "set_unique_names"                                     
#> [707] "set_universal_names"                                  
#> [708] "setdiff.data.frame"                                   
#> [709] "setequal.data.frame"                                  
#> [710] "shift"                                                
#> [711] "shift_c"                                              
#> [712] "shift_slice"                                          
#> [713] "show_query"                                           
#> [714] "show_regroups"                                        
#> [715] "show_time"                                            
#> [716] "signal_warnings"                                      
#> [717] "skip_internal_condition"                              
#> [718] "slice"                                                
#> [719] "slice.data.frame"                                     
#> [720] "slice_"                                               
#> [721] "slice_.data.frame"                                    
#> [722] "slice_.tbl_df"                                        
#> [723] "slice_combine"                                        
#> [724] "slice_eval"                                           
#> [725] "slice_head"                                           
#> [726] "slice_head.data.frame"                                
#> [727] "slice_max"                                            
#> [728] "slice_max.data.frame"                                 
#> [729] "slice_min"                                            
#> [730] "slice_min.data.frame"                                 
#> [731] "slice_rank_idx"                                       
#> [732] "slice_rows"                                           
#> [733] "slice_sample"                                         
#> [734] "slice_sample.data.frame"                              
#> [735] "slice_tail"                                           
#> [736] "slice_tail.data.frame"                                
#> [737] "some"                                                 
#> [738] "sort_key_generator"                                   
#> [739] "sql"                                                  
#> [740] "sql_escape_ident"                                     
#> [741] "sql_escape_string"                                    
#> [742] "sql_join"                                             
#> [743] "sql_select"                                           
#> [744] "sql_semi_join"                                        
#> [745] "sql_set_op"                                           
#> [746] "sql_subquery"                                         
#> [747] "sql_translate_env"                                    
#> [748] "src"                                                  
#> [749] "src_df"                                               
#> [750] "src_local"                                            
#> [751] "src_mysql"                                            
#> [752] "src_postgres"                                         
#> [753] "src_sqlite"                                           
#> [754] "src_tbls"                                             
#> [755] "src_tbls.src_local"                                   
#> [756] "standardise_join_incomplete"                          
#> [757] "standardise_join_no_match"                            
#> [758] "standardise_join_remaining"                           
#> [759] "standardise_join_suffix"                              
#> [760] "stop_dplyr"                                           
#> [761] "stop_input_type"                                      
#> [762] "stop_invalid_dollar_sign"                             
#> [763] "stop_invalid_namespaced_expression"                   
#> [764] "stop_invalid_top_expression"                          
#> [765] "stop_join"                                            
#> [766] "stop_join_by_closest_equal_expression"                
#> [767] "stop_join_by_closest_invalid_expression"              
#> [768] "stop_join_matches_multiple"                           
#> [769] "stop_mask_type"                                       
#> [770] "stop_pick_empty"                                      
#> [771] "str_rep"                                              
#> [772] "strip_pos"                                            
#> [773] "summarise"                                            
#> [774] "summarise.data.frame"                                 
#> [775] "summarise.grouped_df"                                 
#> [776] "summarise.rowwise_df"                                 
#> [777] "summarise_"                                           
#> [778] "summarise_.data.frame"                                
#> [779] "summarise_.tbl_df"                                    
#> [780] "summarise_all"                                        
#> [781] "summarise_at"                                         
#> [782] "summarise_build"                                      
#> [783] "summarise_bullets"                                    
#> [784] "summarise_bullets.dplyr:::summarise_incompatible_size"
#> [785] "summarise_bullets.dplyr:::summarise_mixed_null"       
#> [786] "summarise_bullets.dplyr:::summarise_unsupported_type" 
#> [787] "summarise_cols"                                       
#> [788] "summarise_deprecate_variable_size"                    
#> [789] "summarise_each"                                       
#> [790] "summarise_each_"                                      
#> [791] "summarise_each_impl"                                  
#> [792] "summarise_eval_one"                                   
#> [793] "summarise_if"                                         
#> [794] "summarise_inform"                                     
#> [795] "summarise_verbose"                                    
#> [796] "summarize"                                            
#> [797] "summarize_"                                           
#> [798] "summarize_all"                                        
#> [799] "summarize_at"                                         
#> [800] "summarize_each"                                       
#> [801] "summarize_each_"                                      
#> [802] "summarize_if"                                         
#> [803] "symdiff"                                              
#> [804] "symdiff.data.frame"                                   
#> [805] "symdiff.default"                                      
#> [806] "tally"                                                
#> [807] "tally.data.frame"                                     
#> [808] "tally_"                                               
#> [809] "tally_n"                                              
#> [810] "tbl"                                                  
#> [811] "tbl.DBIConnection"                                    
#> [812] "tbl.src_local"                                        
#> [813] "tbl_at_syms"                                          
#> [814] "tbl_at_vars"                                          
#> [815] "tbl_df"                                               
#> [816] "tbl_if_syms"                                          
#> [817] "tbl_if_vars"                                          
#> [818] "tbl_nongroup_vars"                                    
#> [819] "tbl_ptype"                                            
#> [820] "tbl_ptype.default"                                    
#> [821] "tbl_sum.grouped_df"                                   
#> [822] "tbl_sum.rowwise_df"                                   
#> [823] "tbl_vars"                                             
#> [824] "tbl_vars.data.frame"                                  
#> [825] "tbl_vars_dispatch"                                    
#> [826] "the"                                                  
#> [827] "tick"                                                 
#> [828] "tick_if_needed"                                       
#> [829] "top_frac"                                             
#> [830] "top_n"                                                
#> [831] "top_n_rank"                                           
#> [832] "topic_links"                                          
#> [833] "transmute"                                            
#> [834] "transmute.data.frame"                                 
#> [835] "transmute_"                                           
#> [836] "transmute_.data.frame"                                
#> [837] "transmute_all"                                        
#> [838] "transmute_at"                                         
#> [839] "transmute_if"                                         
#> [840] "transpose"                                            
#> [841] "ungroup"                                              
#> [842] "ungroup.data.frame"                                   
#> [843] "ungroup.grouped_df"                                   
#> [844] "ungroup.rowwise_df"                                   
#> [845] "uninline"                                             
#> [846] "union.data.frame"                                     
#> [847] "union_all"                                            
#> [848] "union_all.data.frame"                                 
#> [849] "union_all.default"                                    
#> [850] "unique_names"                                         
#> [851] "universal_names"                                      
#> [852] "unstructure"                                          
#> [853] "validate_and_split_formula"                           
#> [854] "validate_grouped_df"                                  
#> [855] "validate_recode_default"                              
#> [856] "validate_rowwise_df"                                  
#> [857] "vars"                                                 
#> [858] "vars_select_syms"                                     
#> [859] "vec_as_location_invert"                               
#> [860] "vec_case_match"                                       
#> [861] "vec_case_when"                                        
#> [862] "vec_order_radix"                                      
#> [863] "vec_paste0"                                           
#> [864] "vec_slice2"                                           
#> [865] "vec_type_friendly"                                    
#> [866] "walk"                                                 
#> [867] "warn_dplyr"                                           
#> [868] "warn_filter_across"                                   
#> [869] "warn_filter_data_frame"                               
#> [870] "warn_filter_one_column_matrix"                        
#> [871] "warn_join"                                            
#> [872] "warn_join_cross_by"                                   
#> [873] "warn_join_multiple"                                   
#> [874] "warn_join_multiple_null"                              
#> [875] "warn_text_se"                                         
#> [876] "warn_underscored"                                     
#> [877] "with_groups"                                          
#> [878] "with_no_rlang_infix_labeling"                         
#> [879] "with_order"                                           
#> [880] "wrap"                                                 
#> [881] "wrap_dbplyr_obj"